requirements.txt
discord.py==2.2.2
aiosqlite==0.17.0